권한부여

+++ src/main/java/com/koreaIT/java/am/ArticleDoDeleteServlet.java
@@ -1,19 +1,20 @@
 package com.koreaIT.java.am;
 
+import java.io.IOException;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+import java.util.Map;
+
+import com.koreaIT.java.am.util.DBUtil;
+import com.koreaIT.java.am.util.SecSql;
+
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-import java.util.List;
-import java.util.Map;
-
-import com.koreaIT.java.am.util.DBUtil;
-import com.koreaIT.java.am.util.SecSql;
+import jakarta.servlet.http.HttpSession;
 
 @WebServlet("/article/doDelete")
 public class ArticleDoDeleteServlet extends HttpServlet {
@@ -34,6 +35,39 @@ public class ArticleDoDeleteServlet extends HttpServlet {
 			int id = Integer.parseInt(request.getParameter("id"));
 			
 			SecSql sql = new SecSql();
+			
+			sql.append("SELECT * FROM");
+			sql.append("article");
+			sql.append("WHERE id = ?", id);
+			
+			Map<String, Object> articleMap = DBUtil.selectRow(conn, sql);
+			if(articleMap.isEmpty()) {
+				response.setContentType("text/html; charset=UTF-8");
+				response.getWriter().append(String.format("<script> alert('해당글은 존재하지 않습니다.'); location.replace('../home/main');</script>"));
+				return;
+			}
+			
+			int loginedMemberId = -1;
+			
+			HttpSession session = request.getSession();
+			
+			if(session.getAttribute("loginedMemberId") != null) {
+				loginedMemberId = (int)session.getAttribute("loginedMemberId");
+			}
+			
+			if(loginedMemberId == -1) {
+				response.setContentType("text/html; charset=UTF-8");
+				response.getWriter().append(String.format("<script> alert('로그인 후 이용해주세요.'); location.replace('../member/login');</script>"));
+				return;
+			}
+			
+			if((int)articleMap.get("memberId") != loginedMemberId) {
+				response.setContentType("text/html; charset=UTF-8");
+				response.getWriter().append(String.format("<script> alert('권한이 없습니다.'); location.replace('../home/main');</script>"));
+				return;
+			}
+			
+			sql = new SecSql();
 			
 			sql.append("DELETE FROM");
 			sql.append("article");
